requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
psycopg2-binary==2.9.9